ICE_ID is now a fnv1 hash created from the ontology iri connection of sets to nodes in ASP includes the modeled hirachie now

+++ ice/include/ice/Identifier.h
@@ -19,7 +19,7 @@
 namespace ice
 {
 /** Type definition of the time data type */
-typedef int identifier;
+typedef unsigned long identifier;
 
 //* IDGenerator
 /**
