+++ ice/include/ice/ros/RosCommunication.h
@@ -66,7 +66,7 @@ private:
   ros::Subscriber                       heartbeatSubscriber;            /**< Subscriber on the heartbeat channel */
   ros::Publisher                        coordinationPublisher;          /**< Publisher on the coordination channel */
   ros::Subscriber                       coordinationSubscriber;         /**< Subscriber on the coordination channel */
-  int                                   iceId;                          /**< own ice id */
+  long                                  iceId;                          /**< own ice id */
   bool                                  running;                        /**< True if the worker thread is running, else false */
   std::thread                           worker;                         /**< Worker thread to perform the spin */
   el::Logger*                           _log;                           /**< Logger for communication */
